% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_knapsack.R
\name{benchmark_knapsack}
\alias{benchmark_knapsack}
\title{The \code{benchmark_knapsack} function}
\usage{
benchmark_knapsack(x_list, W_list, name_list = NULL)
}
\arguments{
\item{x_list}{A list of dataframes containing the objects that can go in the knapsacks.
Each dataframe has two variables \code{v} indicating the value of the object, and \code{w} indicating the weight of the object.}

\item{W_list}{A list of numerics indicating the size of the knapsacks. In other words: the maximum weights that the respective knapsacks can support.}

\item{name_list}{An (optional) list of names for the datasets.}
}
\value{
A dataframe with one column each for the brute, dynamic and greedy algorithms. Elements are the optimal value achieved for a particular dataset.
}
\description{
This function runs all three knapsack algorithms (included in this package) on some given dataframes of objects.
}
