% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_knapsack.R
\name{dynamic_knapsack}
\alias{dynamic_knapsack}
\title{The \code{dynamic_knapsack} function}
\source{
\href{https://en.wikipedia.org/wiki/Knapsack_problem#0-1_knapsack_problem}{Dynamic programming algorithm from the wikipedia article on the Knapsack problem}
}
\usage{
dynamic_knapsack(x, W)
}
\arguments{
\item{x}{A dataframe containing the objects that can go in the knapsack.
It has two variables \code{v} indicating the value of the object, and \code{w} indicating the weight of the object. For the dynamic solution, the weights must be positive integers.}

\item{W}{A numeric (must be positive integer) indicating the size of the knapsack. In other words: the maximum weight that the knapsack can support.}
}
\value{
A list of $value: the maximum knapsack value (numeric), and $elements: The rows in \code{x} that achieve this value, in the form of a vector.
}
\description{
This function solves the knapsack problem using a method of dynamic programming.
}
